+++ packages/analytics-js/src/components/core/utilities.ts
@@ -1,3 +1,6 @@
+import { isString } from '@rudderstack/analytics-js-common/utilities/checks';
+import { isValidURL } from '@rudderstack/analytics-js-common/utilities/url';
+
 const dispatchSDKEvent = (event: string): void => {
   const customEvent = new CustomEvent(event, {
     detail: { analyticsInstance: (globalThis as typeof window).rudderanalytics },
@@ -9,4 +12,8 @@ const dispatchSDKEvent = (event: string): void => {
   (globalThis as typeof window).document.dispatchEvent(customEvent);
 };
 
-export { dispatchSDKEvent };
+const isWriteKeyValid = (writeKey: string) => isString(writeKey) && writeKey.trim().length > 0;
+
+const isDataPlaneUrlValid = (dataPlaneUrl: string) => isValidURL(dataPlaneUrl);
+
+export { dispatchSDKEvent, isWriteKeyValid, isDataPlaneUrlValid };