+++ packages/analytics-v1.1/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.48.28](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.48.27...rudder-sdk-js@2.48.28) (2024-11-18)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.0`
 ## [2.48.27](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.48.26...rudder-sdk-js@2.48.27) (2024-11-18)
 
 ### Dependency Updates