+++ packages/loading-scripts/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.0.43](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.42...@rudderstack/analytics-js-loading-scripts@3.0.43) (2024-11-18)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.11.0`
 ## [3.0.42](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.41...@rudderstack/analytics-js-loading-scripts@3.0.42) (2024-11-18)
 
 ### Dependency Updates